FCT Seed fonctionnel

+++ TP3/Web/FlappyBird/Models/Score.cs
@@ -7,7 +7,7 @@ namespace FlappyBird.Models
         public int Id { get; set; }
        
         public int ScoreValue {  get; set; }
-        public decimal TimeInSeconds { get; set; }
+        public double TimeInSeconds { get; set; }
         public DateTime? Date { get; set; }
         public bool IsPublic { get; set; }
