+++ TP3/Web/FlappyBird/Models/ScoreDTO.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
 
         public int ScoreValue { get; set; }
-        public decimal TimeInSeconds { get; set; }
+        public double TimeInSeconds { get; set; }
         public DateTime? Date { get; set; }
         public bool IsPublic { get; set; }
         public string Pseudo { get; set; }
